Revert maven-compiler-plugin configuration and use ASM instead
JmxStarterIntegrationTest should be able to run the TestApplication with
different (including older) JVMs. This means that TestApplication needs
to be compiled with an older target level than the rest of the project.
Different executions of the maven-compiler-plugin with appropriate
inclusion/exclusion settings would do the job. However, M2Eclipse is not
behaving well with that kind of configuration.
As a solution to the problem, the class file TestApplication will be
re-written to a temporary folder with 0x31 (Java 5) as bytecode version.